pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nv-cli"
version = "7.0.0"
description = "NVIDIA AI CLI Tool v7.0 - OpenClaw-inspired multi-agent framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "NVIDIA CLI"}
]
keywords = ["nvidia", "ai", "claude", "openclaw", "agent", "llm", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pathspec>=0.12.0",
    "prompt-toolkit>=3.0.0",
    "sentence-transformers>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
nv = "nv_cli.cli:app"

[project.urls]
Homepage = "https://github.com/nvidia/nv-cli"
Documentation = "https://docs.nvidia.com/nv-cli"

[tool.hatch.build.targets.wheel]
packages = ["nv_cli"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true